anomaly-detection-server/requirements.txt
fastapi
uvicorn
pandas
email-validator
python-dotenv
joblib
scikit-learn
paho-mqtt==1.6.1